sagemaker-api-gateway-caching/requirements.txt
aws_cdk.aws_sagemaker>=1.135.0
aws_cdk.aws_iam>=1.135.0
aws-cdk.core>=1.135.0
aws-cdk.aws-apigateway>=1.135.0
